use the provided exponential backoff constructor
[concourse/concourse#6770]

we missed some default values when we updated the package, so let's just
rely on the constructor to initialize it to some sane value.

what ended up happening was that one of those default values was a magic
number, which caused an inifinite retry since the default initialized
value was not the same.

Signed-off-by: Clara Fu <cfu@pivotal.io>
Co-authored-by: Bohan Chen <bochen@pivotal.io>